refactor: rename wgpu_layer to wgpu_wrapper
