+++ examples/misc/src/main.rs
@@ -1,7 +1,7 @@
 use nalgebra::{Scale3, Translation3, Unit, UnitQuaternion, Vector3};
 use reverie_engine::{
     scene::{EntityIndex, Frame, Scene, SpriteComponent, System, TransformComponent},
-    wgpu_layer::WgpuResource,
+    wgpu_wrapper::WgpuResource,
     Game,
 };
 use winit::event::{ElementState, MouseButton};
