+++ reverie-engine/src/lib.rs
@@ -5,7 +5,7 @@
 mod game;
 pub mod scene;
 pub mod texture;
-pub mod wgpu_layer;
+pub mod wgpu_wrapper;
 mod winit_app;
 
 pub use game::start_engine;
